requirements.txt
asgiref==3.3.1
Django==3.1.7
django-autoslug==1.9.8
django-crispy-forms==1.11.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
